Refactor GovUKTableRadioField into fields.py
